+++ Actions/.Modules/ReadSettings.psm1
@@ -241,6 +241,7 @@ function GetDefaultSettings
         "gitSubmodulesTokenSecretName"                  = "gitSubmodulesToken"
         "shortLivedArtifactsRetentionDays"              = 1  # 0 means use GitHub default
         "reportSuppressedDiagnostics"                   = $false
+        "workflowInputDefaults"                         = @()
     }
 }
 
@@ -590,4 +591,5 @@ function SanitizeWorkflowName {
 }
 
 Export-ModuleMember -Function ReadSettings
+Export-ModuleMember -Function SanitizeWorkflowName
 Export-ModuleMember -Variable ALGoFolderName, ALGoSettingsFile, RepoSettingsFile, CustomTemplateRepoSettingsFile, CustomTemplateProjectSettingsFile